Fix lint

+++ src/backend/src/signer.rs
@@ -97,7 +97,7 @@ async fn cfs_ecdsa_pubkey_of(principal: &Principal) -> Result<Vec<u8>, String> {
     let (ecdsa_key_name, maybe_cfs_canister_id) =
         read_config(|s| (s.ecdsa_key_name.clone(), s.cfs_canister_id));
     // As done in [CFS](https://github.com/dfinity/chain-fusion-signer/blob/26b683c6de9971fdbf7bd4cebc04d427d1753289/src/signer/canister/src/derivation_path.rs#L28)
-    let derivation_path = vec![vec![0 as u8], principal.as_slice().to_vec()];
+    let derivation_path = vec![vec![0_u8], principal.as_slice().to_vec()];
     let cfs_canister_id = maybe_cfs_canister_id.ok_or("Missing CFS canister id")?;
     if let Ok((key,)) = ecdsa_public_key(EcdsaPublicKeyArgument {
         canister_id: Some(cfs_canister_id),
